// React/Site/src/API.ts
export type CreateUserInput = {
  id?: string | null,
  fullName: string,
  firebaseUid: string,
  email: string,
  displayName: string,
  imageUri?: string | null,
  defaultMidiDevice?: string | null,
  purchasedTokens: number,
  remainingTokens: number,
  totalStreams: number,
  totalSeconds: number,
  lastStream?: string | null,
  createdInstanceId?: string | null,
};
// ...
export type ModelStringInput = {
  ne?: string | null,
  eq?: string | null,
  le?: string | null,
  lt?: string | null,
  ge?: string | null,
  gt?: string | null,
  contains?: string | null,
  notContains?: string | null,
  between?: Array< string | null > | null,
  beginsWith?: string | null,
  attributeExists?: boolean | null,
  attributeType?: ModelAttributeTypes | null,
  size?: ModelSizeInput | null,
};

export enum ModelAttributeTypes {
  binary = "binary",
  binarySet = "binarySet",
  bool = "bool",
  list = "list",
  map = "map",
  number = "number",
  numberSet = "numberSet",
  string = "string",
  stringSet = "stringSet",
  _null = "_null",
}


export type ModelSizeInput = {
  ne?: number | null,
  eq?: number | null,
  le?: number | null,
  lt?: number | null,
  ge?: number | null,
  gt?: number | null,
  between?: Array< number | null > | null,
};

export type ModelIDInput = {
  ne?: string | null,
  eq?: string | null,
  le?: string | null,
  lt?: string | null,
  ge?: string | null,
  gt?: string | null,
  contains?: string | null,
  notContains?: string | null,
  between?: Array< string | null > | null,
  beginsWith?: string | null,
  attributeExists?: boolean | null,
  attributeType?: ModelAttributeTypes | null,
  size?: ModelSizeInput | null,
};

export type ModelIntInput = {
  ne?: number | null,
  eq?: number | null,
  le?: number | null,
  lt?: number | null,
  ge?: number | null,
  gt?: number | null,
  between?: Array< number | null > | null,
  attributeExists?: boolean | null,
  attributeType?: ModelAttributeTypes | null,
};

export type User = {
  __typename: "User",
  id: string,
  fullName: string,
  firebaseUid: string,
  email: string,
  displayName: string,
  imageUri?: string | null,
  defaultMidiDevice?: string | null,
  purchasedTokens: number,
  remainingTokens: number,
  totalStreams: number,
  totalSeconds: number,
  lastStream?: string | null,
  createdInstanceId?: string | null,
  hostRoomUser?: ModelHostRoomUserConnection | null,
  createdAt: string,
  updatedAt: string,
};

export type ModelHostRoomUserConnection = {
  __typename: "ModelHostRoomUserConnection",
  items:  Array<HostRoomUser | null >,
  nextToken?: string | null,
};

export type HostRoomUser = {
  __typename: "HostRoomUser",
  id: string,
  userId: string,
  hostRoomId: string,
  lastSeen?: string | null,
  archived: boolean,
  pinned: boolean,
  user?: User | null,
  hostRoom?: HostRoom | null,
  createdAt: string,
  updatedAt: string,
  userHostRoomUserId?: string | null,
  hostRoomHostRoomUsersId?: string | null,
  hostRoomUserUserId?: string | null,
};

export type HostRoom = {
  __typename: "HostRoom",
  id: string,
  name: string,
  desc?: string | null,
  imageUri?: string | null,
  createdUserId: string,
  lastMessageId?: string | null,
  lastMessage?: HostRoomChatMessage | null,
  chatMessages?: ModelHostRoomChatMessageConnection | null,
  hostRoomUsers?: ModelHostRoomUserConnection | null,
  createdAt: string,
  updatedAt: string,
  hostRoomLastMessageId?: string | null,
};
// ...
export type UpdateUserInput = {
  id: string,
  fullName?: string | null,
  firebaseUid?: string | null,
  email?: string | null,
  displayName?: string | null,
  imageUri?: string | null,
  defaultMidiDevice?: string | null,
  purchasedTokens?: number | null,
  remainingTokens?: number | null,
  totalStreams?: number | null,
  totalSeconds?: number | null,
  lastStream?: string | null,
  createdInstanceId?: string | null,
};
// ...
export type ModelUserFilterInput = {
  id?: ModelIDInput | null,
  fullName?: ModelStringInput | null,
  firebaseUid?: ModelIDInput | null,
  email?: ModelStringInput | null,
  displayName?: ModelStringInput | null,
  imageUri?: ModelStringInput | null,
  defaultMidiDevice?: ModelStringInput | null,
  purchasedTokens?: ModelIntInput | null,
  remainingTokens?: ModelIntInput | null,
  totalStreams?: ModelIntInput | null,
  totalSeconds?: ModelIntInput | null,
  lastStream?: ModelStringInput | null,
  createdInstanceId?: ModelStringInput | null,
  and?: Array< ModelUserFilterInput | null > | null,
  or?: Array< ModelUserFilterInput | null > | null,
  not?: ModelUserFilterInput | null,
};

export enum ModelSortDirection {
  ASC = "ASC",
  DESC = "DESC",
}


export type ModelUserConnection = {
  __typename: "ModelUserConnection",
  items:  Array<User | null >,
  nextToken?: string | null,
};
// ...
export type CreateUserMutation = {
  createUser?:  {
    __typename: "User",
    id: string,
    fullName: string,
    firebaseUid: string,
    email: string,
    displayName: string,
    imageUri?: string | null,
    defaultMidiDevice?: string | null,
    purchasedTokens: number,
    remainingTokens: number,
    totalStreams: number,
    totalSeconds: number,
    lastStream?: string | null,
    createdInstanceId?: string | null,
    hostRoomUser?:  {
      __typename: "ModelHostRoomUserConnection",
      items:  Array< {
        __typename: "HostRoomUser",
        id: string,
        userId: string,
        hostRoomId: string,
        lastSeen?: string | null,
        archived: boolean,
        pinned: boolean,
        createdAt: string,
        updatedAt: string,
        userHostRoomUserId?: string | null,
        hostRoomHostRoomUsersId?: string | null,
        hostRoomUserUserId?: string | null,
      } | null >,
      nextToken?: string | null,
    } | null,
    createdAt: string,
    updatedAt: string,
  } | null,
};

export type UpdateUserMutationVariables = {
  input: UpdateUserInput,
  condition?: ModelUserConditionInput | null,
};

export type UpdateUserMutation = {
  updateUser?:  {
    __typename: "User",
    id: string,
    fullName: string,
    firebaseUid: string,
    email: string,
    displayName: string,
    imageUri?: string | null,
    defaultMidiDevice?: string | null,
    purchasedTokens: number,
    remainingTokens: number,
    totalStreams: number,
    totalSeconds: number,
    lastStream?: string | null,
    createdInstanceId?: string | null,
    hostRoomUser?:  {
      __typename: "ModelHostRoomUserConnection",
      items:  Array< {
        __typename: "HostRoomUser",
        id: string,
        userId: string,
        hostRoomId: string,
        lastSeen?: string | null,
        archived: boolean,
        pinned: boolean,
        createdAt: string,
        updatedAt: string,
        userHostRoomUserId?: string | null,
        hostRoomHostRoomUsersId?: string | null,
        hostRoomUserUserId?: string | null,
      } | null >,
      nextToken?: string | null,
    } | null,
    createdAt: string,
    updatedAt: string,
  } | null,
};
// ...
export type ListUsersQuery = {
  listUsers?:  {
    __typename: "ModelUserConnection",
    items:  Array< {
      __typename: "User",
      id: string,
      fullName: string,
      firebaseUid: string,
      email: string,
      displayName: string,
      imageUri?: string | null,
      defaultMidiDevice?: string | null,
      purchasedTokens: number,
      remainingTokens: number,
      totalStreams: number,
      totalSeconds: number,
      lastStream?: string | null,
      createdInstanceId?: string | null,
      hostRoomUser?:  {
        __typename: "ModelHostRoomUserConnection",
        nextToken?: string | null,
      } | null,
      createdAt: string,
      updatedAt: string,
    } | null >,
    nextToken?: string | null,
  } | null,
};